refactor: Move base cards code into own directory
